% file grt/man/subjdemo_2d.Rd
\name{subjdemo_2d}
\alias{subjdemo_2d}
\docType{data}
\title{
Sample dataset of a categorization experiment with 2D stimuli.
}
\description{
A sample two dimensional stimulus set and response data of a hypothetical participant in a two-category classification experiment involving 500 trials.
}
\usage{
subjdemo_2d
}
\format{
  This data frame contains 500 rows and the following columns:
  \describe{
    \item{\code{category}}{
      label of the category to which each stimulus belongs.
    }
    \item{\code{x}}{value on the dimension 'x'}
    \item{\code{y}}{value on the dimension 'y'}
    \item{\code{response}}{
      classification response of a participant.
    }
  }
}
\source{
  Alfonso-Reese, L. A. (2006)
  \emph{General recognition theory of categorization: A MATLAB toolbox}.
  Behavior Research Methods, 38, 579-583.
}
\references{
  Alfonso-Reese, L. A. (2006)
  \emph{General recognition theory of categorization: A MATLAB toolbox}.
  Behavior Research Methods, 38, 579-583.
}
\keyword{datasets}
